+++ src/main/kotlin/me/sivieri/aoc2020/day14/MemoryInput.kt
@@ -2,5 +2,5 @@ package me.sivieri.aoc2020.day14
 
 data class MemoryInput(
     val bitMask: BitMask,
-    val values: List<Pair<Int, Long>>
+    val values: List<Pair<Long, Long>>
 )
